+++ config/email.php
@@ -1,0 +1,27 @@
+<?php
+
+/**
+ * --------------------------------------------------------------------------------
+ * Email SMTP Config
+ * --------------------------------------------------------------------------------
+ */
+
+return [
+    'default' => 'smtp',
+    'smtp' => [
+        'host' => 'smtp.qq.com',
+        'port' => '465',
+        'username' => 'xxxxxxx@qq.com',
+        'password' => 'xxxxxxxxxxxxxx',
+        'hostname' => 'www.shejibiji.com',
+        'fromemailname' => 'shejibiji',
+    ],
+    'pop3' => [
+        'host' => 'smtp.qq.com',
+        'port' => '465',
+        'username' => 'xxxxxxx@qq.com',
+        'password' => 'xxxxxxxxxxxxxx',
+        'hostname' => 'www.shejibiji.com',
+        'fromemailname' => 'shejibiji',
+    ],
+];